ion: ensure prefetch/drain requests are processed in order
Currently when there is an ION prefetch or drain request we are adding
them to the start of the prefetch_list list.

This can cause the requests to be processed in the wrong order, for
example if there is a prefetch request made followed by a drain request
(before the prefetch request has been processed) the the drain request
will be handled first.

Ensure prefetch and drain requests are processed in order by appending
requests to the end of the prefetch_list list

Change-Id: I935a0d69a52267e888e9b19e1e8c0d9bd68e295b
Signed-off-by: Liam Mark <lmark@codeaurora.org>
Signed-off-by: KyuoFoxHuyu <88403766@qq.com>